+++ src/xyz/oihana/schema/constants/traits/business/DocumentsTrait.php
@@ -8,6 +8,7 @@ use xyz\oihana\schema\constants\traits\business\documents\BusinessDocumentLineTr
 use xyz\oihana\schema\constants\traits\business\documents\BusinessDocumentTrait;
 use xyz\oihana\schema\constants\traits\business\documents\DocumentTotalsTrait;
 use xyz\oihana\schema\constants\traits\business\documents\EcoFeeRuleTrait;
+use xyz\oihana\schema\constants\traits\business\documents\InvoiceTrait;
 use xyz\oihana\schema\constants\traits\business\documents\PaymentInstallmentTrait;
 use xyz\oihana\schema\constants\traits\business\documents\PaymentScheduleTrait;
 use xyz\oihana\schema\constants\traits\business\documents\QuoteTrait;
@@ -22,14 +23,15 @@ use xyz\oihana\schema\constants\traits\business\documents\TaxDetailTrait;
  */
 trait DocumentsTrait
 {
-    use AdjustmentTrait          ,
-        AppliedEcoFeeTrait       ,
+    use AdjustmentTrait           ,
+        AppliedEcoFeeTrait        ,
         BusinessDocumentLineTrait ,
-        BusinessDocumentTrait    ,
-        DocumentTotalsTrait      ,
-        EcoFeeRuleTrait          ,
-        PaymentInstallmentTrait  ,
-        PaymentScheduleTrait     ,
-        QuoteTrait               ,
-        TaxDetailTrait           ;
+        BusinessDocumentTrait     ,
+        DocumentTotalsTrait       ,
+        EcoFeeRuleTrait           ,
+        InvoiceTrait              ,
+        PaymentInstallmentTrait   ,
+        PaymentScheduleTrait      ,
+        QuoteTrait                ,
+        TaxDetailTrait            ;
 }